test(core-state): crate backup & empty wallet-repository-clone.test.ts
